[idna] Drop tables module
